feat(widget): add geoLookup method in TimelineProvider
