+++ IVPNClient/Utilities/Extensions/UserDefaultsTodayExtension+Ext.swift
@@ -55,14 +55,14 @@ extension UserDefaults {
         return Date().addingTimeInterval(TimeInterval(-60))
     }
     
-    var geoLookup: GeoLookup? {
+    var geoLookup: GeoLookup {
         if let saved = UserDefaults.shared.object(forKey: UserDefaults.Key.geoLookup) as? Data {
             if let loaded = try? JSONDecoder().decode(GeoLookup.self, from: saved) {
                 return loaded
             }
         }
         
-        return nil
+        return GeoLookup(ipAddress: "", countryCode: "", country: "", city: "", isIvpnServer: false, isp: "", latitude: 0, longitude: 0)
     }
     
 }
