+++ IVPNWidget/Scene/Location/LocationViewModel.swift
@@ -27,17 +27,15 @@ import WidgetKit
 extension LocationView {
     class ViewModel: ObservableObject {
         
+        @Published var model: GeoLookup
         let apiService: APIService
         
         init(apiService: APIService = WidgetAPIService()) {
             self.apiService = apiService
+            self.model = UserDefaults.shared.geoLookup
         }
         
         func getLocation() -> String {
-            guard let model = UserDefaults.shared.geoLookup else {
-                return ""
-            }
-            
             guard !model.city.isEmpty else {
                 return model.country
             }
@@ -46,30 +44,7 @@ extension LocationView {
         }
         
         func getCountryCode() -> String {
-            guard let model = UserDefaults.shared.geoLookup else {
-                return ""
-            }
-            
             return model.countryCode.uppercased()
-        }
-        
-        func geoLookup() {
-            guard Date().timeIntervalSince(UserDefaults.shared.lastWidgetUpdate) > 3 else {
-                return
-            }
-            
-            UserDefaults.shared.set(Date(), forKey: UserDefaults.Key.lastWidgetUpdate)
-            
-            let requestIPv4 = ApiRequestDI(method: .get, endpoint: Config.apiGeoLookup, addressType: .IPv4)
-            apiService.request(requestIPv4) { (result: Result<GeoLookup>) in
-                switch result {
-                case .success(let model):
-                    model.save()
-                    WidgetCenter.shared.reloadTimelines(ofKind: "IVPNWidget")
-                case .failure:
-                    break
-                }
-            }
         }
         
     }
